graph-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
grpcio==1.62.1
grpcio-tools==1.62.1
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-instrumentation-grpc==0.46b0
neo4j==5.28.1
